Refactor: Extract wallet logic from signals.py into wallet_service.py
